[General]: #pragma once all over, for faster builds.

+++ src/baseq2rtxp/svgame/svg_level_locals.h
@@ -22,7 +22,7 @@ struct svg_level_locals_t {
     uint64_t    frameNumber = 0;
     QMTime		time = 0_ms;
 
-    float       gravity = 0.f;
+    double      gravity = 0.;
 
     const cm_entity_t **cm_entities = nullptr;
